+++ Space-Fox.Unity/Assets/Scripts/Meshes/IMesh.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+using UnityEngine;
+
+namespace SpaceFox
+{
+    public interface IMesh
+    {
+        List<Vector3> Vertices { get; }
+        int[] GetTrianglesAsPlainArray();
+    }
+}
