+++ Space-Fox.Unity/Assets/Scripts/Meshes/IMesh.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 13adbae06092d024f888ef9383b2cc96
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
